+++ src/layouts/index.js
@@ -3,28 +3,19 @@ import PropTypes from 'prop-types'
 import Helmet from 'react-helmet'
 
 import Header from '../components/Header'
-import './index.css'
+import 'normalize.css'
 
-const TemplateWrapper = ({ children }) => (
+const TemplateWrapper = ({ data: { site }, children }) => (
   <div>
     <Helmet
-      title="Gatsby Default Starter"
+      title={site.siteMetadata.title}
       meta={[
         { name: 'description', content: 'Sample' },
         { name: 'keywords', content: 'sample, something' },
       ]}
     />
-    <Header />
-    <div
-      style={{
-        margin: '0 auto',
-        maxWidth: 960,
-        padding: '0px 1.0875rem 1.45rem',
-        paddingTop: 0,
-      }}
-    >
-      {children()}
-    </div>
+    <Header title={site.siteMetadata.title} />
+    {children()}
   </div>
 )
 
@@ -33,3 +24,14 @@ TemplateWrapper.propTypes = {
 }
 
 export default TemplateWrapper
+
+export const query = graphql`
+  # inject sitedata into the data-prop
+  query SiteDataQuery {
+    site {
+      siteMetadata {
+        title
+      }
+    }
+  }
+`